LayoutTests/fast/workers/shared-worker-messageevent-source.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Make sure that MessageEvent.source is properly set in connect event.");

startWorker("resources/messageevent-source.js", true);
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
